First commited as before_6121_rev1

+++ org.eclipse.jgit/src/org/eclipse/jgit/merge/ResolveMerger.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (C) 2010, Christian Halstrick <christian.halstrick@sap.com>,
- * Copyright (C) 2010-2012, Matthias Sohn <matthias.sohn@sap.com>
+ * Copyright (C) 2010, Matthias Sohn <matthias.sohn@sap.com>
  * and other copyright owners as documented in the project's IP log.
  *
  * This program and the accompanying materials are made available
@@ -78,6 +78,7 @@ import org.eclipse.jgit.lib.ConfigConstants;
 import org.eclipse.jgit.lib.Constants;
 import org.eclipse.jgit.lib.FileMode;
 import org.eclipse.jgit.lib.ObjectId;
+import org.eclipse.jgit.lib.ObjectInserter;
 import org.eclipse.jgit.lib.ObjectReader;
 import org.eclipse.jgit.lib.Repository;
 import org.eclipse.jgit.treewalk.CanonicalTreeParser;
@@ -126,11 +127,11 @@ public class ResolveMerger extends ThreeWayMerger {
 
 	private Map<String, DirCacheEntry> toBeCheckedOut = new HashMap<String, DirCacheEntry>();
 
-	private List<String> toBeDeleted = new ArrayList<String>();
-
 	private Map<String, MergeResult<? extends Sequence>> mergeResults = new HashMap<String, MergeResult<? extends Sequence>>();
 
 	private Map<String, MergeFailureReason> failingPaths = new HashMap<String, MergeFailureReason>();
+
+	private ObjectInserter oi;
 
 	private boolean enterSubtree;
 
@@ -154,6 +155,7 @@ public class ResolveMerger extends ThreeWayMerger {
 				SupportedAlgorithm.HISTOGRAM);
 		mergeAlgorithm = new MergeAlgorithm(DiffAlgorithm.getAlgorithm(diffAlg));
 		commitNames = new String[] { "BASE", "OURS", "THEIRS" };
+		oi = getObjectInserter();
 		this.inCore = inCore;
 
 		if (inCore) {
@@ -223,8 +225,8 @@ public class ResolveMerger extends ThreeWayMerger {
 				builder = null;
 			}
 
-			if (getUnmergedPaths().isEmpty() && !failed()) {
-				resultTree = dircache.writeTree(getObjectInserter());
+			if (getUnmergedPaths().isEmpty()) {
+				resultTree = dircache.writeTree(oi);
 				return true;
 			} else {
 				resultTree = null;
@@ -242,20 +244,15 @@ public class ResolveMerger extends ThreeWayMerger {
 			for (Map.Entry<String, DirCacheEntry> entry : toBeCheckedOut
 					.entrySet()) {
 				File f = new File(db.getWorkTree(), entry.getKey());
-				createDir(f.getParentFile());
-				DirCacheCheckout.checkoutEntry(db, f, entry.getValue(), r);
+				if (entry.getValue() != null) {
+					createDir(f.getParentFile());
+					DirCacheCheckout.checkoutEntry(db, f, entry.getValue(), r);
+				} else {
+					if (!f.delete())
+						failingPaths.put(entry.getKey(),
+								MergeFailureReason.COULD_NOT_DELETE);
+				}
 				modifiedFiles.add(entry.getKey());
-			}
-			// Iterate in reverse so that "folder/file" is deleted before
-			// "folder". Otherwise this could result in a failing path because
-			// of a non-empty directory, for which delete() would fail.
-			for (int i = toBeDeleted.size() - 1; i >= 0; i--) {
-				String fileName = toBeDeleted.get(i);
-				File f = new File(db.getWorkTree(), fileName);
-				if (!f.delete())
-					failingPaths.put(fileName,
-							MergeFailureReason.COULD_NOT_DELETE);
-				modifiedFiles.add(fileName);
 			}
 		} finally {
 			r.release();
@@ -402,7 +399,7 @@ public class ResolveMerger extends ThreeWayMerger {
 					else {
 						// the preferred version THEIRS has a different mode
 						// than ours. Check it out!
-						if (isWorktreeDirty(work))
+						if (isWorktreeDirty())
 							return false;
 						DirCacheEntry e = add(tw.getRawPath(), theirs,
 								DirCacheEntry.STAGE_0);
@@ -437,7 +434,7 @@ public class ResolveMerger extends ThreeWayMerger {
 			// THEIRS. THEIRS is chosen.
 
 			// Check worktree before checking out THEIRS
-			if (isWorktreeDirty(work))
+			if (isWorktreeDirty())
 				return false;
 			if (nonTree(modeT)) {
 				DirCacheEntry e = add(tw.getRawPath(), theirs,
@@ -448,7 +445,7 @@ public class ResolveMerger extends ThreeWayMerger {
 			} else if (modeT == 0 && modeB != 0) {
 				// we want THEIRS ... but THEIRS contains the deletion of the
 				// file
-				toBeDeleted.add(tw.getPathString());
+				toBeCheckedOut.put(tw.getPathString(), null);
 				return true;
 			}
 		}
@@ -488,7 +485,7 @@ public class ResolveMerger extends ThreeWayMerger {
 
 		if (nonTree(modeO) && nonTree(modeT)) {
 			// Check worktree before modifying files
-			if (isWorktreeDirty(work))
+			if (isWorktreeDirty())
 				return false;
 
 			MergeResult<RawText> result = contentMerge(base, ours, theirs);
@@ -510,7 +507,7 @@ public class ResolveMerger extends ThreeWayMerger {
 				// OURS was deleted checkout THEIRS
 				if (modeO == 0) {
 					// Check worktree before checking out THEIRS
-					if (isWorktreeDirty(work))
+					if (isWorktreeDirty())
 						return false;
 					if (nonTree(modeT)) {
 						if (e != null)
@@ -566,7 +563,7 @@ public class ResolveMerger extends ThreeWayMerger {
 		return isDirty;
 	}
 
-	private boolean isWorktreeDirty(WorkingTreeIterator work) {
+	private boolean isWorktreeDirty() {
 		if (inCore)
 			return false;
 
@@ -574,13 +571,8 @@ public class ResolveMerger extends ThreeWayMerger {
 		final int modeO = tw.getRawMode(T_OURS);
 
 		// Worktree entry has to match ours to be considered clean
-		final boolean isDirty;
-		if (nonTree(modeF))
-			isDirty = work.isModeDifferent(modeO)
-					|| !tw.idEqual(T_FILE, T_OURS);
-		else
-			isDirty = false;
-
+		final boolean isDirty = nonTree(modeF)
+				&& !(modeO == modeF && tw.idEqual(T_FILE, T_OURS));
 		if (isDirty)
 			failingPaths.put(tw.getPathString(),
 					MergeFailureReason.DIRTY_WORKTREE);
@@ -627,8 +619,7 @@ public class ResolveMerger extends ThreeWayMerger {
 			dce.setLength((int) of.length());
 			InputStream is = new FileInputStream(of);
 			try {
-				dce.setObjectId(getObjectInserter().insert(
-				    Constants.OBJ_BLOB, of.length(), is));
+				dce.setObjectId(oi.insert(Constants.OBJ_BLOB, of.length(), is));
 			} finally {
 				is.close();
 				if (inCore)
